Optimized buttons & added to multiview
